requirements.txt
flask
flask-pagedown
flask-wtf
flask-bootstrap
flask-socketio
markdown2
Markdown
google-api-python-client
google-api-support